pyproject.toml
[project]
name = "nda-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "anthropic>=0.89.0",
    "docx-editor>=0.2.1",
    "fastapi>=0.135.3",
    "mammoth>=1.12.0",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.43.0",
]

[project.scripts]
nda-review = "nda_generator.cli:main"
nda-web = "nda_generator.web:serve"

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nda_generator"]

[tool.uv]
package = true
